refactor(settings-dialog): extract gamepad-nav into a part file
Move the gamepad/keyboard navigation slice of
_SystemEmulatorSettingsDialogState (_initializeGamepad, _cleanupGamepad,
_navigateUp/_navigateDown, _previousTab/_nextTab, _handleSelectItem)
verbatim into system_emulator_settings_dialog/gamepad_nav.dart as
`extension _GamepadNav`. First part/extension split of C.

State stays on the host (extensions can't declare fields); the 8
setState calls route through a new host `rebuild()` bridge
(State.setState is @protected and can't be invoked from an extension).
No public-API or behaviour change.
